+++ resources/js/pages/welcome.tsx
@@ -23,7 +23,7 @@ export default function Welcome({ users, paginator }: WelcomeProps) {
 
     const [isSearchLoading, setIsSearchLoading] = useState(false);
 
-    const initialUsersRef = useRef<User[]>(users);
+    const initialUsersRef = useRef<User[]>(paginator.data);
     const initialTotalRef = useRef(paginator.total);
 
     const _users = initialUsersRef.current;
@@ -61,7 +61,7 @@ export default function Welcome({ users, paginator }: WelcomeProps) {
                         <div className="flex-1" />
                         {auth.user ? (
                             <Link
-                                href={route('feed')}
+                                href={route('hunt-line')}
                                 className="inline-block rounded-sm border border-[#19140035] px-5 py-1.5 text-sm leading-normal text-[#1b1b18] hover:border-[#1915014a] dark:border-[#3E3E3A] dark:text-[#EDEDEC] dark:hover:border-[#62605b]"
                             >
                                 Hunt Line